refactored signals

+++ users/signals.py
@@ -1,7 +1,7 @@
 from django.conf import settings
 from django.contrib.auth.models import User
 from django.core.mail import send_mail
-from django.db.models.signals import post_save, post_delete
+from django.db.models.signals import post_save, post_delete, pre_delete
 from django.dispatch import receiver
 
 
@@ -12,7 +12,7 @@ from .ranks import DIVISIONS, RANKS
 '''Creates a profile for the user'''
 @receiver(post_save, sender=User, dispatch_uid="create_profile")
 def CreateProfile(sender, instance, created, **kwargs):
-    if created == True:
+    if created:
         user = instance
         profile = Profile.objects.create(
             user=user,
@@ -35,19 +35,18 @@ def CreateProfile(sender, instance, created, **kwargs):
 '''Updates the user information if the profile is changed.'''
 @receiver(post_save, sender=Profile, dispatch_uid="update_user")
 def UpdateUser(sender, instance, created, **kwargs):
-    profile = instance
-    user = profile.user
-    if created == False:
-        user.first_name = profile.name
-        user.username = profile.username
-        user.email = profile.email
-        user.rank = profile.rank
-        user.division = profile.division
-        if profile.is_coach:
+    user = instance.user
+    if not created:
+        user.first_name = instance.name
+        user.username = instance.username
+        user.email = instance.email
+        user.rank = instance.rank
+        user.division = instance.division
+        if instance.is_coach:
             try:
                 coach_obj, is_coach = Coach.objects.get_or_create(
-                        user = profile,
-                        display_name = profile.username,
+                        user = instance,
+                        display_name = instance.username,
                         headline = "",
                         body = "",
                         discord_link = ""
@@ -58,16 +57,23 @@ def UpdateUser(sender, instance, created, **kwargs):
                 print(e)
                 print(f"Exception was thrown in the {function.__name__}")
         else:
-            Coach.objects.filter(user=profile).delete()
+            Coach.objects.filter(user=instance).delete()
         user.save()
         
 '''Handles Deletion of user if the profile is deleted'''
 @receiver(post_delete, sender=Profile, dispatch_uid="delete_user")
 def DeleteUser(instance, **kwargs):
-        try:
-            user = instance.user
-            user.profile.delete()
-            user.delete()
-        except RecursionError as e:
-            print("Recursion Error in DeleteUser signal.")
-            print(e.message)
+  try:
+      user = instance.user
+      if instance.user:
+        user.profile.delete()
+      else:
+        print(instance)    
+      user.delete()
+  except RecursionError as e:
+      print("Recursion Error in DeleteUser signal.")
+      print(e.message)
+
+# @receiver(pre_delete, sender=Profile, dispatch_uid="test_delete")
+# def TestSig(instance, **kwargs):
+#     print(instance)
